+++ onnxruntime/test/onnx/heap_buffer.h
@@ -4,9 +4,6 @@
 #pragma once
 #include <vector>
 #include <memory>
-namespace onnxruntime {
-struct OrtCallback;
-}
 
 /**
  * A holder for delay freed buffers
@@ -23,9 +20,7 @@ class HeapBuffer {
     buffers_.push_back(p);
     return p;
   }
-  void AddDeleter(onnxruntime::OrtCallback* d);
 
  private:
-  std::vector<onnxruntime::OrtCallback*> deleters_;
   std::vector<void*> buffers_;
 };
