--- onnxruntime/test/perftest/heap_buffer.cc
@@ -1,10 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-#include "heap_buffer.h"
-
-HeapBuffer::~HeapBuffer() {
-  for (void* p : buffers_) {
-    free(p);
-  }
-}
